python-course-5-hw/src/posts/views.py
from rest_framework.decorators import action
from rest_framework.response import Response

class PostViewSet(viewsets.ModelViewSet):
    @action(detail=True, methods=['post'])
    def like(self, request, pk=None):
        post = self.get_object()
        user = request.user
        if user in post.likes.all():
            post.likes.remove(user)
            status = 'unliked'
        else:
            post.likes.add(user)
            status = 'liked'
        return Response({'status': status})