python-course-5-hw/src/requirements/dev.txt
asgiref==3.8.1
Django==5.1.7
djangorestframework==3.16.0
drf-yasg==1.21.10
inflection==0.5.1
packaging==24.2
pillow==11.1.0
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.1.1
